2021_11_25_pybcn_meetup/requirements.txt
Flask==2.0.1
google-cloud-datastore==2.4.0
gunicorn==20.1.0
google-cloud-storage==1.43.0